+++ src/domain.js
@@ -1,9 +1,0 @@
-/**
- * Domain
- * This is the base class with which all domains draw their common
- * behavior. It can also be extended.
- */
-
-export default function Domain () {
-  console.warn('Extending from Domain is deprecated. Instead just use regular class.')
-}