+++ src/merge.js
@@ -1,5 +1,3 @@
-const hasOwn = Object.prototype.hasOwnProperty
-
 /**
  * Merge any number of objects into a provided object
  *
@@ -7,6 +5,9 @@ const hasOwn = Object.prototype.hasOwnProperty
  * @params {...Object} values - Any number of objects to fold properties into the first
  * @return {Object} The first provided value, extended with all others
  */
+
+import hasOwn from './has-own'
+
 export default function merge (subject) {
   for (var i = 1, len = arguments.length; i < len; i++) {
     var next = arguments[i]