+++ src/tag.js
@@ -10,9 +10,9 @@ const FALLBACK = 'microcosm_action'
  * @private
  */
 export default function tag (fn, name) {
-  if (fn == null) {
-    throw new TypeError('Unable to identify ' + fn + ' action. Did you push the correct action?')
-  }
+  console.assert(fn != null,
+                 'Unable to identify ' + fn + ' action. Did you push the',
+                 'correct action?')
 
   if (fn.hasOwnProperty('toString')) {
     return fn